returning status for a ship now working

+++ Rest/src/main/java/ship/ParametersRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ParametersRepository extends CrudRepository<SavedParameters, Long> {
+public interface ParametersRepository extends CrudRepository<SavedParameters, Integer> {
 
     List<SavedParameters> findByRequiredETA(String requiredETA);
 }
